Create current page state and refactor to use it
We move the API call from the axios.get into the initial state of a new state variable
currentPageUrl.

We then refactor the useEffect hook to take currentPageUrl in the get
call as well as passing it as the second parameter in the useEffect
hook. This means whenever currentPageUrl changes, the useEffect will
re-render, hence fetching the new Url