Create and fill next and prev PageUrl components
We create new state variables which are `nextPageUrl` and `prevPageUrl`.
The inital state of these is set to nothing, we then populate them using
values from the API we call in useEffect.

The json response contains variables, next & previous, which contain the
url of the next 20 or previous 20 paginated results.