Create loading text while data is fetching
We create another state variable `loading` which has an initial state of
false. When UseEffect is called (ie we are going to call the API) we
setLoading to true.

Once the call has been sucessful, we set loading back to true again (as
we are no longer loading the data from the api)

We then have a simple if statement that says if loading is true, print
out loading ... to the user