Use axios cancel token to catch old calls
This is so that if a new call is started when an old one hasnt
finished yet, the old call is cancelled. It means we never get old data
comding through.

We do this by making useEffect return a function which gets called
every time useEffect is called again - so we can clean up after the last
call.

Axios has a built in thing called a CancelToken. By creating an empty
function called cancel and then setting it to a CancelToken within axios.get.

We then make useEffect run cancel() every time its called which means it
will cancel any previous calls when it starts a new one